requirements.txt
Flask>=2.0
mysql-connector-python>=8.0
python-dotenv>=1.0
Werkzeug>=2.0
